backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
transformers
tree_sitter
gitpython
openai
requests
